Format all data sets in Postgresql
This code is used to allow for a join between crime and weather data and census units of analysis. First a table is created from the census shape file with hourly units of analysis. Then crime and weather data are joined to this table to create a dataset with hourly weather and crime count observations for each census unit in the city for the 2001 to 2014 period for which crime data is available.